Displayed processes owned by an user.
It will only work for system users as we get all process returned
by `supervisorclt status` and we check agains the user that owns
the process.

We're also getting the current system user running the command by
default. To get a custom user, it is necessary add --no-system or
provide an user with -u or --user.

Refs: #2
Fixes: #6